DTD-3298: Changes Int to BigInt

+++ app/uk/gov/hmrc/timetopayproxy/models/affordablequotes/AffordableQuoteInstalment.scala
@@ -29,7 +29,7 @@ final case class AffordableQuoteInstalment(
   instalmentInterestAccrued: BigDecimal,
   instalmentBalance: BigDecimal,
   debtItemOriginalDueDate: LocalDate,
-  expectedPayment: Int
+  expectedPayment: BigInt
 )
 
 object AffordableQuoteInstalment {
